EnvProxy API changes
Lots of docs, lots of small tweaks to the API. This feels pretty good. See fairly comprehensive list of changes below. Conceptually the interfaces better match the intent of using the objects. I don't love the pattern of having a "base" class and inhering from it. But I also don't want to jump through hoops to guarantee logic and methods are available for all EnvProxy types.

We're currently implementing the prepend and override logic. We're missing append, which I don't think we'll need. We're also missing default, which might be useful, but I'll wait until I actually need it before I try to define an interface for it. 

I'm also moving the classes to their own files as I think this interface is likely good enough to stick around for awhile.

Here's a detailed list of changes

## EnvProxy

- EnvProxy.export method signature changed so that directory kwargs end in *_dir
- EnvProxy.export behavior changed so it writes to the profile d and export file at the same time instead of needing to be called twice with different arguments

## EnvProxy::Base

- EnvProxy::Base#write_layer changed to write all layers at once instead of only writing one named layer at a time. This removes `name:` from the method signature.
- EnvProxy::Base#touches_layer? removed since `write_layer` now writes all layers at once
- EnvProxy::Base#to_export method signature changed from `replace_app_dir:` to `replace:, with:`. This method is now implemented on the base class instead of sub classes
- EnvProxy::Base#value_for_export is now a required implemented method from subclasses
- EnvProxy::Base#layer_key is now implemented by the base class, the interface for changing the layer filename has moved to #layer_env_type
- EnvProxy::Base#layer_env_type is now expected to be implemented by subclasses it can be :append, :overwrite, :default, or :prepend
- Moved to it's own file
- Class documentation with an example
- Every method has documentation with an example

## EnvProxy::Value renamed to EnvProxy::Override

The name of the class better matches the CNB logic. 

- Now implements `layer_env_type` of :override since setting this value is intended to replace other values
- Internal #validate! class added. Validation logic is moved to that method and is now more robust
- #to_export removed since it is now implemented in the base class. Instead `value_for_export` is implemented which is now consumed by EnvProxy::Base#to_export
- Moved to it's own file
- Class documentation with an example
- Every method has documentation


## EnvProxy::Array renamed to EnvProxy::Prepend

The name of the class better matches the CNB logic

- Now implements `layer_env_type` of :prepend since setting this value is intended to prepend existing values
- #to_export removed since it is now implemented in the base class. Instead `value_for_export` is implemented which is now consumed by EnvProxy::Base#to_export
- Moved to it's own file
- Class documentation with an example
- Every method has documentation