Detect and raise on test ENV mutation
When writing tests it's easy to accidentally mutate a value that will affect another test. It's especially easy to mutate an ENV var. This code will run before the suite to record the initial ENV state and then again after. If it differs then you'll see something like this:

```
$ be rspec spec/unit/
............................
An error occurred in an `after(:suite)` hook.
Failure/Error: raise environment_mutated_message

RuntimeError:
  Something mutated the environment on accident.

  Diff:
    ENV['lol'] changed from '' to 'blerg'
# ./spec/spec_helper.rb:45:in `block (2 levels) in <top (required)>'


Finished in 33.37 seconds (files took 0.35687 seconds to load)
28 examples, 0 failures, 1 error occurred outside of examples
```

As you can see from the code, it's not perfect as I WANT to mutate the HEROKU_API_KEY. But I only want to set this value when needed (if a hatchet test is run for example). To accommodate for some flexibility I'm allowing the ability to register env var keys that can be ignored.

What should you do when a test starts failing due to ENV var mutation? When that happens find a way to avoid or clean up the mutation. Alternatively execute the code in its own fork or process to avoid polluting the top level environment.

Note that while this guards against env mutation this code won't stop all types of shared state mutation, there's still: disk access, in-memory code modification (mutating a constant for example), and external IO effects (such as calls to a database or API). So always keep your eyes peeled for leaking state between tests 👁👄👁.