Ensure that Nokogiri compiles against system
Nokogiri is a HTML parser that is a native extension and needs to compile and link to libxml2.

Nokogiri ships with a version of vendor libxml2 and when you run `bundle install` with nokogiri it defaults to using the vendor version.

The stack image of Heroku provides libxml2. It's important that we tell nokogiri to use the stack image's version of libxml2 otherwise when libxml2 receives a security patch via the stack, nokogiri will not receive the patch.

To facilitate this we can set the env var NOKOGIRI_USE_SYSTEM_LIBRARIES=true. While this is tested in the existing Ruby buildpack https://github.com/heroku/heroku-buildpack-ruby/blob/38cb9c1e8e16fd09cdf2086387fe9a1882d920de/spec/hatchet/bugs_spec.rb#L4-L8 the test is not valid as it will not emit a warning even if it does not build with system ruby. This test is more explicit about the behavior we're desiring