Explicit CNB entry point
When specifying processes in a launch, the entrpoint to the container changes:

```
$ docker inspect minimal-heroku-buildpack-ruby-tests:blerg | grep Entrypoint -A2
            "Entrypoint": null,
            "OnBuild": null,
            "Labels": null
--
            "Entrypoint": [
                "/cnb/process/web"
            ],
```

This is discussed here:

https://heroku.slack.com/archives/GJR4HRYKD/p1606763985069200

The entry point can be explicitly passed to the container instead. With `--entrypoint`. Leaving it blank means that the PATH isn't initialized. The default entry point without a launch.toml is `/cnb/lifecycle/launcher`. This PR adds in an explicit entry point to  `docker run` commands.