Fix pack tests
The Ruby buildpack needs a Ruby binary to "bootstrap" its own execution. With v2 this is done at release time and when being used via git dynamically in `bin/compile`.

That same behavior was ported over to this buildpack, however the directory we write to cannot be in the same directory as the buildpack itself. This causes an error:

```
      [INFO] Discovering process types
       [INFO] Procfile declares types -> web
       ERROR: failed to build: exit status 1

       stderr: mkdir: cannot create directory '/cnb/buildpacks/heroku_ruby/0.0.1/vendor/ruby': Permission denied
       ERROR: failed to build: executing lifecycle: failed with status code: 145
```

## Why didn't this break before

Described in buildpacks/pack#1322 the only reason builds were previously working is that another image using the Ruby identifier and the same version were previously installed and they had the files we needed.

## Test Fix

This PR changes moves the behavior of copying binaries out of `bin/compile` for the CNB interface and instead into the `build.sh` interface which the Cutlass testing tool understands https://github.com/heroku/cutlass/.

When cutlass calls `build.sh` it will vendor in the appropriate binaries (if they don't already exist) and then copy everything to the `target` directory.

## Deploy fix

The same process needs to happen before buildpacks are deployed. This buildpack is currently not deployed, only the "legacy" buildpack is github.com/heroku/heroku-buildpack-ruby.

For the v2 deploy process the `publish.Vendor` in the `buildpack.toml` is understood by Heroku's platform/deploy-process.

For CNB, I'm honestly not sure how those files are getting vendored. It actually looks like they're not:

```
$ cd /tmp
$ cat Dockerfile
FROM alpine:3 AS base

FROM public.ecr.aws/heroku-buildpacks/heroku-ruby-buildpack:0.1.3 as buildpack
COPY --from=base / /
```

Then:

```
$ docker build -t inspectbuildpack .
$ docker run -it --rm inspectbuildpack sh
~$ ls cnb/buildpacks/heroku_ruby/0.1.3/vendor/
dotenv.rb     lpxc.rb       syck_hack.rb  toml.rb
```

I'm not sure how this is working unless people are using it without pack and it's manually downloading the binaries every time.

Doing the same trick with `0.1.0` it looks like it's never been vendored:

```
/ # ls cnb/buildpacks/heroku_ruby/0.1.0/vendor/
dotenv.rb     lpxc.rb       syck_hack.rb  toml.rb
```

## Tests

With the introduction of a `build.sh` we can no longer execute tests in parallel in the same way. For more information: https://salesforce.quip.com/veWYASMAibmN.