Fix bundle install invalidation logic
The previous code installed the `bundler` gem to a different layer's path. This meant uninstalling the old gem was a requirement.

After moving bundler to install in it's own layer, the invalidation can happen at the layer level instead. This is less confusing and means that we might not have to re-download bundler even when the Ruby version changes if the bundler version did not change.