Add output to branches in existing_layer_strategy
Problem: When a user sees a run of the buildpack they only have access to what is printed to the output. If two different runs have the exact same output but one fails and one passes it will be confusing for the user and difficult for the maintainer to know what is happening.

Solution: When branching (such as in  existing_layer_strategy ) output a message to the user to indicate when no action is taken. For example when `ExistingLayerStrategy::Keep` is applied. Also output when a surprising/rare event happens such as clearing a cache due to  stack change. A user and maintainer should be able to follow the branching flow of a buildpack based only on the log output. Not just what is happening, but why is it happening.