Initial struct for parsing `gem list` output
We need a way to determine the contents of the Gemfile.lock. Previously when the buildpack was written in Ruby it could require the bundler code and use the built in lockfile parser.

This code works by instead parsing the contents of `bundle list` to obtain a name and version. 

This code does not currently call `bundle list` but is only the Struct for it.

The other alternative would be shelling out to Ruby, formatting as json and then parsing the json. Something like this will get you access:

```
$ time ruby -rbundler -rjson -e "parser = Bundler::LockfileParser.new(File.read('./Gemfile.lock')); hash = parser.specs.each_with_object({}) {|spec, hash| hash[spec.name] = spec.version; puts JSON.dump(hash) }"
```

This method is very fast taking 0.15 seconds and can be done without the gems installed. It could in theory be run async with other code such as `bundle install`. One possible downside is that this method relies on internal bundler structure, however that's no different than the existing buildpack. Another risk is any deprecations or warnings that accidentally go to stdout will ruin the deserialization.

The `bundle list` method is slower taking 0.3 seconds on CodeTriage and requires gems be installed. It could not be run in parallel with `bundle install`. The upside is the `bundle list` command has been around for a long time and the format hasn't changed. I don't know if the time/cost changes with lots of gems.

```
$ time bundle list
Gems included by the bundle:
  * rake (13.0.6)
Use `bundle info` to print more detailed information about a gem
bundle list  0.23s user 0.11s system 94% cpu 0.367 total
$ time ruby -rbundler -rjson -e "parser = Bundler::LockfileParser.new(File.read('./Gemfile.lock')); hash = parser.specs.each_with_object({}) {|spec, hash| hash[spec.name] = spec.version; puts JSON.dump(hash) }"
{"rake":"13.0.6"}
ruby -rbundler -rjson -e   0.11s user 0.05s system 92% cpu 0.177 total
```

We could change the implementation later. It's a good optimization, but for now we need to move forward.