Implement GemVersion parsing
This GemVersion struct approximates (but does not perfectly replicate) the parsing behavior of Rubygem's `Gem::Version` class. 

Effectively this code works by breaking apart a version string then scanning until it hits a part that contains a non-digit character. Anything left over is an "extra" part of the semver.

Next up is to implement version comparisons to allow this struct to be PartialOrd.