ShellCommand
Introduces a struct `ShellCommand` that wraps process::Command so we can centralize feature development. It also allows us to define our own interface which has the benefit of being able to reduce mistakes. For example: When using `Command` directly it's easy to forget to apply an `env` to the command even though with libcnb this must be done every time. Now the methods that invoke the command require an `Env` so the programmer is reminded to call it.

There's two main methods, one for running command silently and one for running the command and streaming it's output to the user as it runs. Both capture stderr/stdout so they can be used later.

I wish that we had a test asserting this behavior. In Ruby I would introduce an `io` field to the class and write to it instead of using `println!`. I tried using a `Box<dyn Write>` field but this complains when passed to another thread. In the long term I would still like to do something like that, for now it's good enough.