Clean up EnvCommand
## Move code to a lib

Allows us to run doc tests against it.

## Move error logic from boolean attribute to a function. Previously `allow_non_zero_exit` could be set to skip returning an Err. Instead we are now storing a custom function that the end user can use to instead do the same thing if they wish or anything else they want to in the closure/function.

## Status code error from std::io::Error

It turns out that `std::process::Command` returns some failures to run the command as an `std::io::Error`. For example if you try to execute a command that does not exist you get an Err. I previously thought this was for low level operating system type errors, seeing that they contain useful information I decided to treat all `std::io::Error` as a status code failure. This gives the expected behavior where a trying to execute a command that does not exist returns a consistent error.