Refactor Rake assets cache usage
The Prior implementation of caching `public/assets` and `tmp/cache/assets` was gnarly. This refactors the layer creation and cache reading/writing logic into a struct to DRY up the code.

It also opens the door to get fancier (if we desire). For instance: We could record the internal state (i.e. whether or not the cache has been written or not) and then on drop, panic if the cache was created but not stored.

Alternatively the old idea for the cache interface was a block where forgetting to write is impossible https://github.com/heroku/buildpacks-ruby/blob/d526a49f81becaf571329a1adf5fff0668a6b99a/lib/heroku_buildpack_ruby/cache_copy.rb.

Overall I like centralizing the logic for layer creation and disk manipulation. I don't love that this struct depends on `BuildContext` and `LayerName`. However, passing in a layer doesn't really reduce complexity, just moves it around. Ultimately I don't think the buildpack author (me) cares how something is being cached, so I like that this abstracts that away, however it is leaky.

We could perhaps dynamically create a LayerName based off of the passed in app directory to be cached. But it still has to know about context.

I would like to be able to "unit" test this, but the reliance on context and layer name means I'll have to do something agressive like introduce my own build context factory.

The other meh thing about this code is I'm unwrapping. There are conditions where the files to be copied will fail, for instance, if that directory doesn't exist. It's good enough for now. However in the future either we want to guard against the most common cases and/or actually turn these into a proper Result returning function.