Make test deterministic
Storing mtime of a file has system limitations   based on precision. Even though file `a` is created before file `b` it might happen so fast that they have the same `mtime` in which case the ordering will be random. 

This change forces different mtimes on files and should be deterministic.