Remove double output when streaming
Right now when you stream something via `EnvCommand` and it fails you get the error output in the original stream, then again when the error is displayed. This makes for a lot of duplicate output, especially in the case of `rake assets:precompile` where it's not uncommon for there to be a hundred+ lines. 

This commit adds an internal switch to the error result struct. When set it will not include the stdout or stderr in the display or debug.