BUNDLE_WITHOUT is now default instead of prepend
It makes more sense to give users control over their BUNDLE_WITHOUT value instead of letting them only prepend to it. They may want to remove environments to exclude, not simply add them. 

In addition to that change I restructured the code so the value is passed in. This would theoretically allow us to re-use this `bundle::install` code in another buildpack if we wanted to install dependencies for test (purely hypothetical at this point).