Move util functions
The functions in `util` were only being used by one file. I don't love having something that sounds like a junk drawer in my project otherwise it tends to accumulate junk. 

If we need to download other things, then we can extract this logic out to commons and find a better name.

I also flattened the 3 enums so we don't have to map twice in the code. This will have to be undone if we want to extract this logic, but let's build that bridge when we cross it.

Errors need to be revisited across the board. Cleaning this up for the short term.