Move SECRET_KEY_BASE to store.toml
The metadata store in individual layers is not considered a durable store, it is a cache and the application should behave correctly when it is missing. 

If SECRET_KEY_BASE changes, then all visitors to the website will be logged off. Which is not great.

This change moves the default generated value to `store.toml` which is the intended durable store and should have stronger guarantees for persisting between deploys.