Remove All panics from InAppDirCache interface
## InternalLayerError

While unlikely, it's still possible that the layer will return an error. Although it's an internal error, we should still give the user the option to handle it if they want.

I don't want to make CacheError generic, so I converted the error to a debug statement which is the only thing B::Error is guaranteed to support. It should give enough information to the programmer even if it doesn't look very nice.

Other existing panics:

- Glob error in least_recently_used_files_above_limit_from_path

## InternalBadGlobError

Sure it's an internal error, but maybe someone wants to catch it, IDK. Don't panic. Propagates the original error.

## MtimeUnsupportedOS

Same as above, don't panic in a library, propagates the original error in this case.