Fix bug in InAppDirCache introduced in refactor
The way we're building a new vec to allow us to avoid the panic meant that we were no longer mutating the `files` vec in the `sort_by`. Running the tests confirmed it was broken.

At the end of the day we need 3 values, size, mtime and the path of the file for this operation. To get these three values we must call two failable operations `p.metadata()` and `p.metadata().unwrap().mtime()` this caused the need to use `and_then` and some nested logic.

Instead, I introduced a new struct to hold all three values and gave it a constructor that returns the error we need which makes the code much cleaner.