Update EnvCommand#stream to use libherokubuildpack
In the process of removing panics I realized that Outcome is almost the exact same struct as command::Output which is standard, so I switched to that instead. To ease the transition I introduced a trait that implements lossy getters for stdout and stderr on Output.

This gives us a more consistent interface that never panics.