Introduce AppCacheCollection
When multiple directories need to be cached together there's a lot of repetitive code. This is an extraction from the rake_assets_precompile step where two directories need to be cached.

The one thing not extracted or abstracted is logging. For now this meets the buildpack's needs. If someone wants similar, but different functionality then it wouldn't be hard to duplicate the code and extend.

Instead of premature abstraction, I want to get this into the wild to see how people might want to use or modify this code.