Refactor bundle install and skip based on cache conditions
This commit first gets rid of the layer `bundle_configure_env` on disk that was written in setup_bundler (moved to download_bundler). That logic is moved into a new layer: BundleInstallLayer. Additionally the GemsPathLayer is removed. Its logic is being ported over to BundleInstallLayer which implements both Create and Update. 

In addition to that lovely refactor I introduced a new feature: conditionally run `bundle install` only when the Gemfile, Gemfile.lock or user env (i.e. context.platform.env()) has been changed.

The upside of this is fewer layers and increased performance on builds with no changes to the Gemfile.lock. The downside of this is that one layer is now huge (about 400 lines).

One other possible downside is usability: Previously people may have expected the output of the bundle install values to show them what was on disk for debugging. We could store the contents of the last bundle install and output it, or emit it some other way. Or it might not actually be that helpful in debugging and it might not need to be added back.

Another potential problem is that if programmers update the LayerEnv chain, it won't be picked up until the next time the developer deploys with a changed Gemfile or Gemfile.lock. I think this is reasonable as it would be a setting designed to affect `bundle install` and would be picked up eventually. Though care should still be taken.