Update Bundle Install Layer
Split up logic states between keeping cache (or not) and comparing diffs.

This way there is never a `ExistingStrategy::Keep`. Any changes to `LayerEnv` will be applied on every run.

Essentially instead of one state machine, now we have two:

- Keep the cache (or not)
- Run the command (or not)

I wrote this in such as way that we could later make a new layer that implements `Layer` something like: `impl LayerWithDigest` it would need:

- run_on_diff(&self)
- build_metadata(&self)
- build_layer_env(&self)
- Self::force_digest (could default to returning None)
- Self::existing_cache_strategy (could default to CacheStrategy::KeepAndRun)

Also I think I can generalize BundleDigest to a more generic BuildLayerDigest. Instead of properties we can maintain a hashmap where the keys are user consumable and the values contain the Shas. 

Make an enum type for Env and for File and provide a builder. One should accept a Platform so we can encourage people not digesting based off of current env accidentally (as it is guaranteed to change for each execution).

We would need a separate cache state for when keys don't match. Possibly one for when values don't match. 

Another option could be to reduce DiffState-s to return its own enum. Basically, Same or different. Then that could hold another enum something like DifferentDiffState::Forced(message) and MismatchedKeys, DifferentValues, Missing.