Move BundleDigest to commons::metadata_digest
This provides a generalized interface for any other buildpacks that want to do a similar comparison.

Right now MetadataDigest can track platform env and files. It's designed so that all fields are `Option` so if we want to add more in the future we can. 

The general idea is unchanged. Build a digest object, store it in metadata. Then on subsequent builds digest the same sources and use that information skip or not.

One added bit of complexity of a general solution versus the original is we no longer have a pre-defined list of files that we want to digest.

This means in addition to the case where files are changed, we have a new case: Different files are inspected. Perhaps someone started tracking only the Gemfile and then later they want to track Gemfile and Gemfile.lock. 

When there's a mismatch like that, I'm reporting back all the files.