Introduce alternative to EnvCommand
The `EnvCommand` was designed when there was no built in way to stream stdout & stderr to the user and also when I was not as experienced with Rust.

The `commons::cmd` takes a few things that were nice about EnvCommand and modularizes them so they can be used piece-meal.

Core goals:

- Talk to the user, includes display functions to render commands to the user
- Consistent errors, Errors need to include maximum information including command and stdout/stderr. Good display by default.
- Consistent experience mapping user versus OS errors.