Use formatdoc! instead of format! in errors (#129)
* Use formatdoc! instead of format!

We don't want to emit the extra whitespace here. Use `formatdoc!` macro instead of `format!` (by mistake).

This would be a hard one to catch. I wonder if there's a way to detect/warn on it somehow. Maybe with clippy and a custom rule?

* Add version to changelog and document

It's unintuitive why the first release was 2.0.0 so I added a note.

* Update README