Replace NameAndThenExt with CmdMapExt
The `name_and_then` function was weird. I can admit that. This uses a more familiar pattern with `map`. I chose not to use `map` because the error you get if you forget to include the trait extension tells you that `Iterator.map` isn't implemented on `Command`. Plus Manuel said he thought it would be weird having `map` on something that isn't a <I don't remember what the word he used was but basically a container struct>.

I think the code is a little busier with an extra newline, but ultimately easier to reason about. I'm happy enough with this to merge.