Introduce new metrics buildpack
The classic `heroku/metrics` buildpack sets up a statsd daemon that the dyno can use to send Heroku language specific metrics. For Ruby the biggest one is Puma Pool Capacity.

This code will download the metrics client and put it on the path, however there's an issue affecting our ability to actually boot a daemon.

## Context

Classic buildpacks (Heroku v2) have acces to a `profile.d` directory they can use to write scripts which execute at boot time. This was largely needed to set build time environment variables because the directory structure of build and runtime was different. It was eventually used by the `heroku/metrics` buildpack to download an run a metrics daemon in the background.

Cloud Native Buildpacks (CNB) removed support for bash in an effort to simplify the spec and allow buildpacks to target extremely lightweight distributions that do not include `bash` (for example if you are running a pure JVM application, then there's no runtime need to have bash on the system, removing it reduces the image size and can be viewed as a reduction in security surface area).

As part of this work to remove bash from the spec the `profile.d`-like interface was removed from the CNB spec.

## Problem

Without a spec supported way to execute a script at startup there is no way for a buildpack to provide a background daemon. 

## Possible solutions

- TBD