Add RAILS_ENV=production to `rake -P` errors
It's common that when debugging developers will see an error from a command then try to reproduce the issue locally. In the case of a rails app RAILS_ENV can produce wildly different behavior. When reproducing an error they must use the same RAILS_ENV as their build target. By adding it to the task name we encourage this behavior.

- Rename error something better "RakeError" is generic CannotDetectRake is more specific
- Fix visibility issue. It should not be `pub`