[Breaking] Introduce `build_output::fmt::Url` and test Error output
Previously url was represented as an `Option<String>`. This posed a problem as the first URL I wanted to link to didn't really vibe with the error text. I changed it so that developers can provide their own label.

In this process I added tests which required that I strip control codes, and clean up the trailing spaces and un-needed newlines with bangs on them. New error output:

```
! ERROR: Error installing Ruby
!
! Could not install the detected Ruby version. Ensure that you're using a supported
! ruby version and try again.
!
! For more information, refer to the following documentation:
! https://devcenter.heroku.com/articles/ruby-support#ruby-versions

Debug information:

Could not create file: You do not have sufficient permissions to access this file: /path/to/file
```

I'm debating if we should do something similar with an enum for `debug_details` to give flexibility over representation. I don't need it yet though, so I'll hold off.