Refactor diagnostic commands into fun_run feature
There are times when a disk operation or command fails and we want to gather more information. To support this case I added a new struct CmdErrorDiagnostics that accepts an io or command error. I've marked this interface as experimental and subject to change. I'm not 100% on the interface but I like the direction.

As a note: I'm feeling like `fun_run` is missing some (optional) primitives, such as a struct that couples a name with a command, or a result that returns an output coupled with a name. When this interface is revisited I want to revisit that possibility as well.