Add fun_run `NamedOutput`, `ResultNameExt`, CmdErrorDiagnostics
- NamedOutput: Mentioned in a prior commit that there's a lot of annoying gymnastics that have to happen to keep track of a command's name when using `Result<Output, CmdError>`. In this case `CmdError` can be converted into an `Output` but an `Output` cannot be converted into a `CmdError` (because it needs a "name" String). By introducing this struct a lot of problems with the end user experience go away. This is a breaking change as several type signatures changed.

Further I added a convenience function (via trait extension) to convert `Result<Output, std::io::Error>` into a `Result<NamedOutput, std::io::Error>`. The idea is to get this result as quickly as possible and then it becomes much easier to manipulate the data (versus previously we were constantly tracking a separate name variable and needed to do several chained functions on `Result` to do the most common of tasks. All of those functions are still available if you want it, but this is

- CmdErrorDiagnostics: A pattern I use is to run a command to gather additional details when a command errors, I also used this to run diagnostics on a disk operation failure. To help here I introduced a generic type `ErrorDiagnostics<E>` and `CmdErrorDiagnostics` which is `ErrorDiagnostics<CmdError>`. Because `Result<NamedOutput, CmdError>` holds all the data needed to represent any possible output state of a command execution we don't need an additional internal struct.

I've labeled `ResultNameExt`, `CmdErrorDiagnostics`, and `ErrorDiagnostics<E>` as "experimental". I feel confident that `NamedOutput` will stay, and should have been added a long time ago.