Refactor ErrorInto into ErrorBuilder
It love it. Much better. It allows you to gradually build up an error output instead of having to format it in one go and be worried about getting newline spacing right.

Each part is treated as its own paragraph with double newline separating paragraphs. Because we track all the output and only build it at the end, we can determine when to insert a `!\n` or a `\n`.

I want to make this generic and use it for all output types:

- Error (done)
- Warning
- Important

I'm a little stuck on naming though and I don't have any warnings right now (other than some inline "help" comments). The goal is to minimize surprises and maximize flexibility while still giving enough context to the underlying output framework to confidently format inputs.

I still wish I had better formatting for `DebugDetail` case of CommandError's. It makes me think this is wanting an implementation as a common interface.

It's getting a little fractally complicated as I think about it so I'm going to try to move forward without solving that problem (as I don't currently have it yet).

I do want to try to tackle the CmdError case as I feel that's really common. I also want to try to output all possible errors as a OneOff binary for debugging purposes.