Introduce PrintControl
Previously the decision whether an output needs to be `println` or `print` had to be made up front, then if you use `print` you have to remember to always end in a newline later. This structure gives us the ability to set a default and then yield control to the caller. That way they could either continue printing on the same line or finalize it and start printing on their own line.