Integration test output API
In Ruby I build CLI output APIs around an injectable IO object like. like this:

```
class PrintFormatter
  attr_reader: io

  def initialize(io: STDOUT)
    @io = io
  end

  def puts(thing)
    @io.puts("yolo: #{thing}")
  end
end

let printer = PrintFormatter.new(
  io: StringIO.new
)
```

You can then access and assert against the contents in a test.

In Rust that would require otherwise stateless functions to take shared mutable state and it seems bad to mess with the mutability requirements just for testing.

Since I "own" the entirety of `build_output` I'm using a single purpose global to capture the contents of what's being logged in test only.

I looked into doing something like this for the `log` crate (logging facade) but there's a number of edge cases you have to deal with and it's ultimately not what that crate was intended to do. This very focused logic is minimal and allows me to test methods together, instead of only in isolation.