Fix log capturing thread safety
While the prior log testing mechanism was thread safe in the technical sense, it wrote all data from all threads to the same string. This worked because we only have a single test asserting on this information, but when I forced other tests to write to the same logger the test failed.

The solution I introduced here tracks content based on thread id. It needs to be injectable as well because we have some logic like `LiveTimingInline` that prints dots to the background in a separate thread. As we own all the printing code we wish to test, this is a trivial contortion and doesn't change the public API.

I also made sure all structs and enums are derive Debug.