Hardcode agentmon url, improve caching
Agentmon hasn't seen a release in 6 years. We can hardcode the url for now.

When the URL doesn't change we can keep the last download around. I always want to update the scripts on disk so then any changes to the files will be picked up, so we'll always call update or create, and never keep.

We could get fancy and use MetadataDigest to cache those too, but writing them is pretty much instantaneous so I don't think it buys us anything.

This change allows us to remove the `cached` dependency.