Preserve long command for debugging
Previously when the command errored the message would only show `gem install bundler --version <version>` however that could lead to weird error messages. For example if `--install-dir` is removed as an argument to the command the message would make no sense.

This change preserves the full command in the case of errors but shows the short command for inline installation.