Introduce living style guide
TODO: Warnings and errors look off based on whitespace before/after the block. This is a symptom of not knowing that came before. I think we can fix this by introducing an AnnounceLogger that must be used to gain access to error/warning/important.

TODO: The `step_and` should really be `step` while the current `step` should be `step_mut`.
TODO: Make error, warning, important chainable.