Introduce fat::HELP and fat::DEBUG_INFO constants
It's easier to work with and makes more sense. I tried using lazy_static and the `colorize` function, but for some reason it told me that my constant didn't implement display. Possibly a static that is `&'static str` is considered a thing that is different than a const `&'static str` but possibly I did something weird in the implementation.

Add docs for these as well as `section_log` in the `print_style_guide`