Remove panic when joining logger
When returning a StreamLogger it's possible for the user to leak an Arc<Mutex<W>> reference which will cause the `Arc::try_unwrap` to fail (if there is more than one single reference).

While not ideal, we can recover from this scenario by continuing on and using the Arc<Mutex<W>> instead of unwrapping it. This will allow us to avoid a possible runtime panic.

The downside is there could be two things emitting to the UI at the same time. If that's the case it's the developers responsibility to notice and investigate.

Basically: Favor not panic the program when we don't have to due to logging logic.

The runtime panics can be caused by:

- A poisoned mutex. We could possibly recover but it's new territory https://blog.logrocket.com/understanding-handling-rust-mutex-poisoning/. For now I think if a mutex is poisoned that propagating the panic makes sense.
- Closed writer: I don't know when this can happen. Presumably someone could pass in a writer and then close it or the system could error or a file could be deleted etc. In that case we could possibly fallback to generating a new stdout and warn. I'm not sure which is better