Failing test for desired warning formatting
Warnings, errors, and information blocks are different from the rest of the other formatting in that they should have some vertical whitespace around them.

The logic for this is easy isolation but difficult when combined.

## Newline front & back

For example, we could add a newline before and after a block and you would get:

```

! It's too crowded here

```

But if you issue two warnings back to back you would get:

```

! It's too crowded here

! It's too crowded here

```

## Newline front

You can partially fix this by only adding a newline to the front only:

```

! It's too crowded here

! It's too crowded here
```

But if the next thing to print isn't a warning it will be missing a bottom newline

## Newline back

```
! It's too crowded here

! It's too crowded here

```

That's not right either

## Possible fix

I believe that for this to work correctly in all the cases we must store state.

### Save last thing on warn/error/important

We could add an option/boolean/whatever to BuildLog to have it store whether or not the last thing that printed was a block and then adjust appropriately. Seems workable but strangely coupled.

### Delayed log

We could save blocks and defer them until the next non-block call (or drop) and then emit them all at once. I don't hate that.

### Logging state

Introduce an `AnnounceLogger` state or something so when you call `error(self: Box<dyn SectionLogger>) -> Box<dyn AnnounceLogger>`. This state transition could be where we insert the newlines i.e. if you call error on an AnnounceLogger it knows to not add an extra starting newline.

The challenge here is getting back to the original state. This is not valid rust code:

```
trait AnnounceLogger {
	fn error(self: Box<Self>, s: &str) -> Box<dyn Self> // i.e. how can we get back to SectionLogger, StartedLogger
}
```

Now there are only two logger states we would need to return to so we could implement each of them as a separate trait. It's not ideal but it's also not horrible.