Fix warning/important/error formatting using trait
As described in the commit that introduced the failing test, this commit implements the "Logging state" trait approach.

To emit a warning/error/important to the output we must first `announce()`.

As I'm typing this up I realize that there's the case where you can `announce().end_announce()` we could patch this up by introducing an intermediate state that only has error, warn, important on it.

I don't like the duplicate code and duplicate trait definition, but the end result seems like a good API.