Refactor data inside of BuildLog
Sometimes we pass data that's meant for only a specific log lifecycle (like `TimedStepLogger` while other times it's meant to be propagated throughout the entire build. Moving this data to a purpose defined struct makes this more clear and makes it easier to add additional things later (if we so desire).

Ultimately I would like a "warn later" feature. I'm thinking that we can preserve various warnings in a vector. The question is if we want to pass it through the lifecycle of the BuildLog or if we want to put it somewhere as a globally mutable state.

# Warn later

## Pass in Build log

Would not be able to add a `warn_later` in `section_log` functions as those don't have access to prior BuildLog state.

## Global mutable

We can mutate from within `section_log` functions, but well...global mutable state is bad. It might cause headaches

## CNB feature

Instead of baking this in at the Rust level we could find a way to expose a folder that all build packs can write to.  Originally I was thinking we could have a `setup_delayed_warn` and a `print_delayed_warn` buildpack pair but that only works if all build packs inbetween are successful.

This makes me feel like a "run this buildpack even if the ones before it fail" could be a useful feature (or possibly a builder extension spec that would allow heroku to implement this as a vendor specific feature, or possibly a CNB first tier feature).

## Analysis

Given this I think the global mutable state is the most promising path forwards.